+++ web/app/index.html
@@ -44,7 +44,13 @@
 
             <hr>
             <h2>Surveys form directive:</h2>
-            <div surveys-form form-structure="form.structure"></div>
+            <h5>GET URL:</h5>
+            <input type="text" ng-model="form.getUrl">
+            <h5>POST URL:</h5>
+            <input type="text" ng-model="form.postUrl">
+            <hr>
+
+            <div surveys-form get-url="form.getUrl" post-url="form.postUrl"></div>
 
         </section>
     </body>
