+++ web/app/scripts/test-app/test.es.js
@@ -18,10 +18,12 @@
             }).then((res) => {
                 const form = res.data.OK.response_data;
 
+
                 this.fields = FieldsetParserService.parseFields(form);
                 this.model = FieldsetParserService.parseModel(form);
 
-                this.structure = res.data.OK.response_data;
+                this.getUrl = 'http://127.0.0.1:8000/forms/1'
+                this.postUrl = 'http://127.0.0.1:8000/forms/1/respond/test-user'
             });
 
         }
