+++ web/app/templates/incuna-surveys/fields/free-text.html
@@ -1,3 +1,3 @@
 <div drf-form-field="to.fieldOptions" class="text" field-id="options.key">
-    <input type="text" class="text-input" id="{{ options.key }}" ng-model="model[to.fieldSetIndex].answers[options.key]">
+    <input type="text" class="text-input" id="{{ options.key }}" ng-model="model[to.fieldSetIndex].answers[options.key]" ng-required="to.fieldOptions.required">
 </div>
