+++ web/app/templates/incuna-surveys/fields/radio.html
@@ -1,6 +1,6 @@
 <div drf-form field="to.fieldOptions" class="radio">
     <div ng-repeat="choice in to.choices" class="checkable radio">
-        <input type="radio" id="{{ options.key }}-{{ $index }}" ng-value="$index" ng-model="model[to.fieldSetIndex].answers[options.key]">
+        <input type="radio" id="{{ options.key }}-{{ $index }}" ng-value="$index" ng-model="model[to.fieldSetIndex].answers[options.key]" ng-required="to.fieldOptions.required">
         <label for="{{ options.key }}-{{ $index }}">{{ choice }}</label>
     </div>
 </div>
