+++ web/app/templates/incuna-surveys/survey-form.html
@@ -1,0 +1,5 @@
+<form ng-submit="submit()">
+    {{ submit }}
+    <formly-form model="model" fields="fields"></formly-form>
+    <input type="submit" id="submit" value="Submit" />
+</form>
